+++ codegen/src/worktable/generator/table.rs
@@ -32,6 +32,7 @@ impl Generator {
             }
         };
 
+        let primary_index_type = &self.columns.primary_keys.1;
         let iter_with = Self::gen_iter_with(row_type);
         let iter_with_async = Self::gen_iter_with_async(row_type);
         let select_executor = self.gen_select_executor();
@@ -94,6 +95,7 @@ impl Generator {
                     WorkTable<
                         #row_type,
                         #pk_type,
+                        #primary_index_type<#pk_type, Link>,
                         #index_type,
                         <#pk_type as TablePrimaryKey>::Generator,
                         #inner_const_name
@@ -111,6 +113,7 @@ impl Generator {
                     WorkTable<
                         #row_type,
                         #pk_type,
+                        #primary_index_type<#pk_type, Link>,
                         #index_type
                     >
                     #persist_type_part
@@ -140,7 +143,7 @@ impl Generator {
                     let pk = row.get_primary_key();
                     let need_to_update = {
                         let guard = Guard::new();
-                        if let Some(_) = self.0.pk_map.peek(&pk, &guard) {
+                        if let Some(_) = TableIndex::peek(&self.0.pk_map, &pk) {
                             true
                         } else {
                             false
@@ -244,8 +247,7 @@ impl Generator {
                         #lit => {
                             let mut limit = q.params.limit.unwrap_or(usize::MAX);
                             let mut offset = q.params.offset.unwrap_or(0);
-                            let guard = Guard::new();
-                            let mut iter = self.0.indexes.#idx_name.iter(&guard);
+                            let mut iter = TableIndex::iter(&self.0.indexes.#idx_name);
                             let mut rows = vec![];
 
                             while let Some((_, l)) = iter.next() {
@@ -273,8 +275,7 @@ impl Generator {
                         #lit => {
                             let mut limit = q.params.limit.unwrap_or(usize::MAX);
                             let mut offset = q.params.offset.unwrap_or(0);
-                            let guard = Guard::new();
-                            let mut iter = self.0.indexes.#idx_name.iter(&guard);
+                            let mut iter = TableIndex::iter(&self.0.indexes.#idx_name);
                             let mut rows = vec![];
 
                             while let Some((_, links)) = iter.next() {
@@ -317,7 +318,7 @@ impl Generator {
                         let mut limit = q.params.limit.unwrap_or(usize::MAX);
                         let mut offset = q.params.offset.unwrap_or(0);
                         let guard = Guard::new();
-                        let mut iter = self.0.pk_map.iter(&guard);
+                        let mut iter = TableIndex::iter(&self.0.pk_map);
                         let mut rows = vec![];
 
                         while let Some((_, l)) = iter.next() {
@@ -397,8 +398,8 @@ impl Generator {
         Ok(quote! {
             pub fn #fn_name(&self, by: #type_) -> Option<#row_ident> {
                 let guard = Guard::new();
-                let link = self.0.indexes.#field_ident.peek(&by, &guard)?;
-                self.0.data.select(*link).ok()
+                let link = TableIndex::peek(&self.0.indexes.#field_ident, &by)?;
+                self.0.data.select(link).ok()
             }
         })
     }
@@ -418,9 +419,7 @@ impl Generator {
         Ok(quote! {
             pub fn #fn_name(&self, by: #type_) -> core::result::Result<SelectResult<#row_ident, Self>, WorkTableError> {
                 let rows = {
-                    let guard = Guard::new();
-                    self.0.indexes.#field_ident
-                        .peek(&by, &guard)
+                    TableIndex::peek(&self.0.indexes.#field_ident, &by)
                         .ok_or(WorkTableError::NotFound)?
                         .iter()
                         .map(|l| *l.as_ref())
@@ -439,7 +438,7 @@ impl Generator {
             pub fn iter_with<F: Fn(#row) -> core::result::Result<(), WorkTableError>>(&self, f: F) -> core::result::Result<(), WorkTableError> {
                 let first = {
                     let guard = Guard::new();
-                    self.0.pk_map.iter(&guard).next().map(|(k, v)| (k.clone(), *v))
+                    TableIndex::iter(&self.0.pk_map).next().map(|(k, v)| (k.clone(), *v))
                 };
                 let Some((mut k, link)) = first else {
                     return Ok(())
@@ -452,7 +451,7 @@ impl Generator {
                 while !ind {
                     let next = {
                         let guard = Guard::new();
-                        let mut iter = self.0.pk_map.range(k.clone().., &guard);
+                        let mut iter = TableIndex::range(&self.0.pk_map, k.clone()..);
                         let next = iter.next().map(|(k, v)| (k.clone(), *v)).filter(|(key, _)| key != &k);
                         if next.is_some() {
                             next
@@ -479,7 +478,7 @@ impl Generator {
             pub async fn iter_with_async<F: Fn(#row) -> Fut , Fut: std::future::Future<Output = core::result::Result<(), WorkTableError>>>(&self, f: F) ->core::result::Result<(), WorkTableError> {
                 let first = {
                     let guard = Guard::new();
-                    self.0.pk_map.iter(&guard).next().map(|(k, v)| (k.clone(), *v))
+                    TableIndex::iter(&self.0.pk_map).next().map(|(k, v)| (k.clone(), *v))
                 };
                 let Some((mut k, link)) = first else {
                     return Ok(())
@@ -492,7 +491,7 @@ impl Generator {
                 while !ind {
                     let next = {
                         let guard = Guard::new();
-                        let mut iter = self.0.pk_map.range(k.clone().., &guard);
+                        let mut iter = TableIndex::range(&self.0.pk_map, k.clone()..);
                         let next = iter.next().map(|(k, v)| (k.clone(), *v)).filter(|(key, _)| key != &k);
                         if next.is_some() {
                             next