+++ src/lib.rs
@@ -24,8 +24,13 @@ pub mod prelude {
     pub use crate::in_memory::{ArchivedRow, Data, DataPages, RowWrapper, StorableRow};
     pub use crate::lock::LockMap;
     pub use crate::primary_key::{PrimaryKeyGenerator, PrimaryKeyGeneratorState, TablePrimaryKey};
-    use crate::table;
-    pub use crate::{lock::Lock, TableIndex, TableRow, WorkTable, WorkTableError};
+    pub use crate::table::select::{
+        Order, SelectQueryBuilder, SelectQueryExecutor, SelectResult, SelectResultExecutor,
+    };
+    pub use crate::{
+        lock::Lock, IndexSet, KeyValue, TableIndex, TableRow, TableSecondaryIndex, WorkTable,
+        WorkTableError,
+    };
     pub use data_bucket::{
         align, map_data_pages_to_general, map_index_pages_to_general, map_tree_index,
         map_unique_tree_index, parse_data_page, parse_page, persist_page, DataPage, GeneralHeader,
@@ -33,11 +38,9 @@ pub mod prelude {
         SizeMeasurable, SizeMeasure, SpaceInfoData, DATA_VERSION, GENERAL_HEADER_SIZE,
         INNER_PAGE_SIZE, PAGE_SIZE,
     };
+
     pub use derive_more::{From, Into};
     pub use lockfree::set::Set as LockFreeSet;
     pub use scc::{ebr::Guard, tree_index::TreeIndex};
-    pub use table::select::{
-        Order, SelectQueryBuilder, SelectQueryExecutor, SelectResult, SelectResultExecutor,
-    };
     pub use worktable_codegen::{PersistIndex, PersistTable};
 }