+++ tests/worktable/mod.rs
@@ -2,6 +2,7 @@ mod array;
 mod base;
 mod config;
 mod custom_pk;
+mod index_type;
 mod option;
 mod tuple_primary_key;
 mod uuid;